微调

+++ fetchDataAndDownloadImage/downloadImageToServer.js
@@ -18,19 +18,42 @@ module.exports.downloadImageToServer = function ({ imageUrl, imageName }) {
       });
       res.on("end", function () {
         console.log("已收到1张完整图片");
-        fs.writeFile(
-          path.join(imageDirPath, imageName),
-          imgData,
-          "binary",
-          function (err) {
+        if (!fs.existsSync(imageDirPath)) {
+          fs.mkdir(imageDirPath, function (err) {
             if (err) {
               console.log(err);
-              reject("服务器端保存1张图片文件已失败");
+              reject("创建图片文件夹失败");
             } else {
-              resolve("服务器端保存1张图片文件已成功");
+              fs.writeFile(
+                path.join(imageDirPath, imageName),
+                imgData,
+                "binary",
+                function (err) {
+                  if (err) {
+                    console.log(err);
+                    reject("服务器端保存1张图片文件已失败");
+                  } else {
+                    resolve("服务器端保存1张图片文件已成功");
+                  }
+                }
+              );
             }
-          }
-        );
+          });
+        } else {
+          fs.writeFile(
+            path.join(imageDirPath, imageName),
+            imgData,
+            "binary",
+            function (err) {
+              if (err) {
+                console.log(err);
+                reject("服务器端保存1张图片文件已失败");
+              } else {
+                resolve("服务器端保存1张图片文件已成功");
+              }
+            }
+          );
+        }
       });
       res.on("error", function (err) {
         reject("服务器端收到图片响应时发生错误");
